+++ packages/firebase_dynamic_links/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+* **Breaking change**. Migrate from the deprecated original Android Support
+  Library to AndroidX. This shouldn't result in any functional changes, but it
+  requires any Android apps using this plugin to [also
+  migrate](https://developer.android.com/jetpack/androidx/migrate) if they're
+  using the original support library.
+
 ## 0.1.1
 
 * Update example to create a clickable and copyable link.